Fix bug

+++ ImageProcessor.py
@@ -10,6 +10,7 @@ from tkinter import Tk, filedialog
 from pdf2image import convert_from_path
 import os
 import shutil
+import sys
 
 JSON_TEMPLATE = "./templ/template.json"
 OUTPUT_PATH = './output'
@@ -20,7 +21,7 @@ HEADER_SOURCE_IMG = "/horizontal_part_1.png"
 FOOTER_SOURCE_IMG = "/horizontal_part_4.png"
 
 class ImageProcessor:
-    def __init__(self, target_path, template_path, output_aligned_path, crop_output_dir, open_window):
+    def __init__(self, target_path, template_path, output_aligned_path, crop_output_dir, open_window, delete_output):
         self.target_path = target_path
         self.template_path = template_path
         self.output_aligned_path = output_aligned_path
@@ -28,13 +29,14 @@ class ImageProcessor:
         self.version = "after_2018"
         self.open_window = open_window
         self.converted_image_path = None  # path to converted PNG if PDF
-        if os.path.exists(OUTPUT_PATH):
+        if os.path.exists(OUTPUT_PATH) and delete_output == "true":
             shutil.rmtree(OUTPUT_PATH)
             print(f"Deleted folder: {OUTPUT_PATH}")
         else:
             print(f"Folder does not exist: {OUTPUT_PATH}")
         # Recreate it
-        os.makedirs(OUTPUT_PATH)
+        if not os.path.exists(OUTPUT_PATH):
+            os.makedirs(OUTPUT_PATH)
 
     def extract_header(self):
         json_output_path = OUTPUT_PATH + "/" + UNPROCESSED_KK_JSON
@@ -60,12 +62,18 @@ class ImageProcessor:
         )
         if not self.target_path:
             print("No file selected.")
+            sys.exit(0)
         else:
             print(f"Selected file: {self.target_path}")
 
     def convert_pdf_to_png(self, pdf_path):
+        # Get the path of the current script
+        current_dir = os.path.dirname(__file__)
+
+        # Point to the 'bin' directory inside the 'poppler' folder
+        poppler_path = os.path.join(current_dir, "poppler", "bin")
         print("Converting PDF to PNG...")
-        pages = convert_from_path(pdf_path, dpi=200)
+        pages = convert_from_path(pdf_path, dpi=200, poppler_path=poppler_path)
         if not pages:
             print("PDF has no pages.")
             return None
@@ -94,12 +102,14 @@ class ImageProcessor:
                 target = cv2.imread(img_path)
                 if target is None:
                     print("Failed to load the image.")
+                    sys.exit(0)
                 else:
                     print("Image loaded successfully.")
             else:
                 print("No image to load.")
         else:
             print("Run cancelled: no file selected.")
+            sys.exit(0)
 
         target = cv2.resize(target, (template.shape[1], template.shape[0]))
         h_img, w_img = target.shape[:2]
